refactor(userOrdersActions,Repository) : corrected the naming
